# api-docs.json
{"openapi":"3.1.0","info":{"title":"My API","description":"API documentation for my application","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Account","description":"Account API to fetch or create accounts"},{"name":"Transaction","description":"Transactions API to create and fetch"},{"name":"Transfer","description":"API's to help transfer balance from one account into another"},{"name":"Customer","description":"Customer API to create customers"}],"paths":{"/api/transfers/transferEvent":{"post":{"tags":["Transfer"],"operationId":"transferBalanceBetweenAccountsEvent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionMessage"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/Transfer"}}}}}}},"/api/transactions/transfer":{"post":{"tags":["Transaction"],"operationId":"transferBalanceBetweenAccounts","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionMessage"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}}}}},"/api/customers/create":{"post":{"tags":["Customer"],"operationId":"createCustomerWithInitialBalance","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/Customer"}}}}}}},"/api/accounts/create/{customerId}":{"post":{"tags":["Account"],"operationId":"createAdditonalAccountsForCustomer","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/Account"}}}}}}},"/api/transfers/{transferId}":{"get":{"tags":["Transfer"],"operationId":"fetchAccountById","parameters":[{"name":"transferId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/Transfer"}}}}}}},"/api/transactions/all":{"get":{"tags":["Transaction"],"operationId":"fetchTransactions","responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}}}}},"/api/transactions/account/{accountId}":{"get":{"tags":["Transaction"],"operationId":"fetchTransactionsByAccount","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}}}}},"/api/customers/{customerId}":{"get":{"tags":["Customer"],"operationId":"fetchCustomerById","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/Customer"}}}}}}},"/api/customers/all":{"get":{"tags":["Customer"],"operationId":"fetchCustomers","responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Customer"}}}}}}}},"/api/accounts/{accountId}":{"get":{"tags":["Account"],"operationId":"fetchAccountById_1","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/Account"}}}}}}},"/api/accounts/all":{"get":{"tags":["Account"],"operationId":"fetchAccounts","responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"}}}}}}}},"/api/accounts/all/{customerId}":{"get":{"tags":["Account"],"operationId":"fetchAccountsByCustomer","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"}}}}}}}}},"components":{"schemas":{"TransactionMessage":{"type":"object","properties":{"transferId":{"type":"string"},"fromAccountId":{"type":"string"},"toAccountId":{"type":"string"},"transactionAmount":{"type":"number","format":"double"}}},"Transfer":{"type":"object","properties":{"id":{"type":"string"},"fromAccountId":{"type":"string"},"toAccountId":{"type":"string"},"transactionAmount":{"type":"number","format":"double"},"transactionIds":{"type":"array","items":{"type":"string"}},"transferStatus":{"type":"string","enum":["OPEN","PROCESSING","COMPLETED"]},"created":{"type":"string","format":"date-time"}}},"Transaction":{"type":"object","properties":{"id":{"type":"string"},"fromAccountId":{"type":"string"},"toAccountId":{"type":"string"},"transactionAmount":{"type":"number","format":"double"},"accountId":{"type":"string"},"transactionType":{"type":"string","enum":["INITIAL_DEPOSIT","TRANSFER_FROM","TRANSFER_TO"]},"created":{"type":"string","format":"date-time"}}},"Customer":{"type":"object","properties":{"givenName":{"type":"string"},"familyName":{"type":"string"},"emailAddress":{"type":"string"}}},"Account":{"type":"object","properties":{"id":{"type":"string"},"accountBalance":{"type":"number","format":"double"},"dateDeposited":{"type":"string","format":"date-time"},"customer":{"$ref":"#/components/schemas/Customer"},"lastUpdated":{"type":"string","format":"date-time"}}}}}}